Update error handling

+++ views/VerificationView.py
@@ -14,20 +14,18 @@ class VerificationView(View):
     @discord.ui.button(label="Access the server!", style=discord.ButtonStyle.green, custom_id='verify_button')
     async def on_callback(self, button: discord.ui.Button, interaction: discord.Interaction):
         guild = interaction.guild
-        guildConfig = GeneralUtils.getConfig('guild')
-        member = interaction.user
 
         await interaction.response.defer(ephemeral=True)
         await asyncio.sleep(0.2)  # Thinking...
-
-        additional_roles = None
         try:
+            guildConfig = GeneralUtils.getConfig('guild')
+            if not guildConfig:
+                raise Exception("Guild config not found")
+            member = interaction.user
+
             additional_roles = OnboardingRoleDAO.getOnboardingRolesByMember(str(member.id))
             OnboardingRoleDAO.deleteOnboardingRolesByMember(str(member.id))
-        except (Exception) as error:
-            print(error)
-            await interaction.followup.send("Error during verification. Contact bot developer or server admin", ephemeral=True)
-        else:
+
             if additional_roles:
                 for role in additional_roles:
                     add_role = discord.utils.get(guild.roles, id=int(role.roleID))
@@ -35,14 +33,17 @@ class VerificationView(View):
 
             onboarding_role = discord.utils.get(
                 guild.roles, id=int(guildConfig['onboarding_role_id']))
-            if onboarding_role:
-                await member.remove_roles(onboarding_role)
-            else:
-                await interaction.followup.send("Role not found. Contact bot developer or server admin", ephemeral=True)
+            if not onboarding_role:
+                raise Exception("ONBOARDING_ROLE_ID not found in Guild config.")
+            
+            await member.remove_roles(onboarding_role)
 
             member_role = discord.utils.get(guild.roles, id=int(guildConfig['member_role_id']))
-            if member_role:
-                await member.add_roles(member_role)
-                await interaction.followup.send(f"You now have the {member_role.name} role! Enjoy your stay in our server!!", ephemeral=True)
-            else:
-                await interaction.followup.send("Role not found. Contact bot developer or server admin", ephemeral=True)
+            if not member_role:
+                raise Exception("MEMBER_ROLE_ID not found in Guild config.")
+
+            await member.add_roles(member_role)
+            await interaction.followup.send(f"You now have the {member_role.name} role! Enjoy your stay in our server!!", ephemeral=True)
+        except Exception as error:
+            print(error)
+            await interaction.followup.send("Error during verification. Contact the bot developer or server admin.", ephemeral=True)
